refactored header files and libs
