+++ PoissonSimulator/alglib/alglibinternal.cpp
