+++ PoissonSimulator/alglib/alglibmisc.cpp
