+++ PoissonSimulator/alglib/alglibmisc.h
