+++ PoissonSimulator/alglib/ap.cpp
