+++ PoissonSimulator/alglib/dataanalysis.cpp
