+++ PoissonSimulator/alglib/dataanalysis.h
