+++ PoissonSimulator/alglib/diffequations.cpp
