+++ PoissonSimulator/alglib/diffequations.h
