+++ PoissonSimulator/alglib/fasttransforms.cpp
