+++ PoissonSimulator/alglib/fasttransforms.h
