+++ PoissonSimulator/alglib/optimization.cpp
