+++ PoissonSimulator/alglib/optimization.h
