+++ PoissonSimulator/alglib/solvers.cpp
