+++ PoissonSimulator/alglib/specialfunctions.cpp
