+++ PoissonSimulator/alglib/specialfunctions.h
