+++ PoissonSimulator/alglib/statistics.cpp
