--- PoissonSimulator/stdafx.cpp
@@ -1,8 +1,0 @@
-// stdafx.cpp: ???????? ????, ?????????? ?????? ??????????? ?????????? ??????
-// Simulator.pch ????? ??????????????????? ??????????
-// stdafx.obj ????? ????????? ?????????????? ????????????????? ???????? ? ????
-
-#include "stdafx.h"
-
-// TODO: ?????????? ?????? ?? ????? ??????????? ?????????????? ????????? ? ????? STDAFX.H
-// , ? ?? ? ?????? ?????
