+++ PoissonSimulator/stdafx.h
@@ -1,6 +1,6 @@
-// stdafx.h: ?????????? ???? ??? ??????????? ????????? ?????????? ??????
-// ??? ?????????? ?????? ??? ??????????? ???????, ??????? ????? ????????????, ??
-// ?? ????? ??????????
+// stdafx.h: включаемый файл дл€ стандартных системных включаемых файлов
+// или включаемых файлов дл€ конкретного проекта, которые часто используютс€, но
+// не часто измен€ютс€
 //
 
 #pragma once
@@ -8,7 +8,7 @@
 #ifndef STDAFX_H
 #define STDAFX_H
 
-#include <cstdio>
+#include <stdio.h>
 #include <functional>
 #include <list>
 #include <chrono>
@@ -18,11 +18,13 @@
 #include <fstream>
 #define _USE_MATH_DEFINES 
 #include <math.h>
+#include <tuple>
+#include <iomanip> 
 
 #include <boost/random.hpp>
 #include <boost/random/exponential_distribution.hpp>
 #include <boost/math/tools/roots.hpp>
 #include "exprtk.hpp"
-#include "interpolation.h"
+#include "alglib/interpolation.h"
 
 #endif
